[refactor] #49 리뷰반영 수정
