+++ app/src/main/java/com/dpm/sixpack/main/MainActivity.kt
@@ -55,9 +55,7 @@ class MainActivity : ComponentActivity() {
                     LocalTimeZone provides currentTimeZone,
                 ) {
                     SixpackTheme(isDebug = BuildConfig.DEBUG) {
-                        MainScreen(
-                            appState = appState,
-                        )
+                        MainScreen(appState = appState)
                     }
                 }
             }
