+++ presentation/src/main/java/com/dpm/sixpack/presentation/routes/freind/components/FriendsLazyColumn.kt
@@ -1,6 +1,7 @@
 package com.dpm.sixpack.presentation.routes.freind.components
 
 import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
@@ -9,7 +10,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.dpm.sixpack.presentation.common.components.preview.DoRunPreviewWrapper
-import com.dpm.sixpack.presentation.routes.freind.contract.FriendItem
+import com.dpm.sixpack.presentation.common.model.FriendItem
 
 @Composable
 fun FriendsLazyColumn(
@@ -18,8 +19,7 @@ fun FriendsLazyColumn(
     onCheerClick: (Long) -> Unit = {},
 ) {
     LazyColumn(
-        modifier = modifier.fillMaxWidth(),
-        // 아이템 사이의 간격
+        modifier = modifier,
         verticalArrangement = Arrangement.spacedBy(8.dp),
     ) {
         // 친구 목록 아이템들을 표시
