+++ presentation/src/main/java/com/dpm/sixpack/presentation/routes/freind/contract/FriendUiState.kt
@@ -1,6 +1,7 @@
 package com.dpm.sixpack.presentation.routes.freind.contract
 
 import com.dpm.sixpack.presentation.common.base.UiState
+import com.dpm.sixpack.presentation.common.model.FriendItem
 import kotlinx.parcelize.Parcelize
 
 @Parcelize
